moving functions aroung broke rna sampler. sampler, abstract molecules and rna forgi are tested and ok
